refactor: simplify configurations
Old style feature modules were shared between all consumers, and they
hard-coded the option values of their features.  If one consumer wanted
to set a different value, the module provided an option for that.  This
resulted in the outside module, which implemented a feature, providing a
set of options, and our feature module, which was meant to set that set
of options to values we wanted, providing an intersecting option set.
This defeated the purpose of having that feature module in the first
place; we could just set the options of the outside module instead.

This commit duplicates these feature modules for each host, system, and
user, making them simpler and more specific for each configuration.

This commit also enables hosts, systems, and users in the flake by
importing them directly, rather than using options to enable them.

 * Import inputs in the configurations that use them, not in flake.nix.
 * Import modules and overlays in flake, where they can be accessed
   from anywhere.