feat: remove protondrive
Protondrive works great on rclone, at least until the system is
upgraded.  After upgrading however, protondrive starts returning
"invalid credentials" errors and the whole logging in process must be
performed all over again.  This is fucking annoying and breaks the
whole point of declarative setups.  Not to mention using rclone might be
potentially hazardous to the wellbeing of Proton accounts, seeing that
Proton does not support / is actively discouraging the use of rclone.

Move to using the web app, until a better alternative is found.